app-grid-button: Add Uninstall action
gnome-software supports an --uninstall <id> invocation, which navigates
directly to the details page for the app and pops the prompt to
uninstall that app.

Unfortunately, this invocation differs from --details in that it
requires a fully-formed AppStream "data id", rather than the much
simpler desktop ID (which we already have from GIO GAppInfo)

The newly introduced PhoshMetainfoCache is used to look this data ID up
based on a application's desktop ID.

Since the cache might be slow to populate, we bind its readiness to the
enablement of the newly introduced Uninstall SimpleAction in the
popover.

Guido and I were concerned that all of these individual popovers (one is
instantiated for every app in the grid) might be a performance issue,
especially now that the enabled property is being bound to other
objects, but Guido indicated that his testing revealed no significant
problems here.

Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1600>