shell: Allow overview to be hidden
This commit introduces a PhoshShell::overview-visible property, which is
bound to the visibility of the PhoshHome that the shell constructs.

This property isn't used in phosh-session and probably never will be,
it's for custom PhoshShell subtypes, such as phrog, which will use it
when displaying a "first-run" application in the system greeter view.

Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1601>