home: Bind background visibility
Instead of always showing it, bind it to the visibility of home.
This ensures they stay in sync.

Since `visible` is false in the constructor this is a slight behavior
change as the background will now become visible a bit later.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1604>