layersurface: Allow to set stacking before the surface is mapped
Catch up with values set prior to mapping the surface on map.

So far this resulted in errors but allowing it makes things more
flexible as we don't need to bother when the surface gets mapped.

We take a weak ref on the target surface so we can notice when it goes
away. Otherwise we would try to bind a non existent target surface when
the target surface goes away before the surface itself gets ever mapped.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1604>