background: Fill background with primary color
This makes sure we have a proper background in the overview in
case the selected image has transparency.

In case of a missing image it ensures we fall back to black so the
overview always gets a non transparent background.

Closes: https://gitlab.gnome.org/World/Phosh/phosh/-/issues/1173
Closes: https://gitlab.gnome.org/World/Phosh/phosh/-/issues/1169
Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1607>