background: Use background color for the primary output's background
The primary output's background determines the color behind the top and
bottom bars when they are folded (since the actual bars are
transparent). So use the @phosh_bg_color CSS value for these areas
rather then the primary-color gsetting.

This ensures the seamless fading from the opened overview to the black
top and bottom panel works even for situations where the primary-color
gsetting is not '#000000'.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1607>