connectivity-manager: Make notification persistent
We keep the notification around unless the SSID of the network changes
or we get connectivity again. Close it as soon as we get connectivity
again or the SSID changes.

The banner will close after 20s (rather than 10s atm) so it is a bit
easier to catch.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1617>