build: Omit phosh_tool_headers from phosh_tool_sources
Similar to the previous commit, we don't want to pollute the
phosh_tool_sources variable with headers. This way the tool sources can
be fed to g-ir-scanner without causing unintended introspection output.

It's worth noting that, for now, the phosh_tool_headers variable *is*
being fed into g-ir-scanner. This is because the work to reduce libphosh
surface area hasn't happened yet. That happens in later commits in this
series.

Signed-off-by: Sam Day <me@samcday.com>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1623>