build: Only add phosh_*_headers to phosh_headers
Again, this continues the theme of the previous commits. We don't want
unexpected headers polluting the list of files that are fed into the
generation of the libphosh GIR.

As with the previous commit, even though we've isolated these headers
for the purpose of omitting them from GIR generation step, they're
explicitly added for the moment to ensure that this commit continues to
successfully build the GIR in the same way it previously did.

Signed-off-by: Sam Day <me@samcday.com>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1623>