build: Keep track of public dbus sources/headers
When codegenning the dbus proxies, we keep track of the files that
should be included in the public API of libphosh.

The newly introduced libphosh_generated_dbus_* variables are not
being used yet, but they will be when the GIR is updated to use a much
smaller subset of files for introspection.

Signed-off-by: Sam Day <me@samcday.com>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1623>