build: Dramatically reduce the size of libphosh public API
The dozen or so commits prior to this one in this series were all
working towards this point. The GIR and libphosh.h are reworked to only
include a very small set of headers for the current minimum viable
public API (based on existing usage in libphosh-rs examples and phrog).

The GIR has been shrunk from ~70k lines to ~2.3k, and the number of
files installed to /usr/include/libphosh-N.NN has reduced from ~200 down
to ~20.

Signed-off-by: Sam Day <me@samcday.com>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1623>