treewide: Remove various docstring symbol links to internal classes
The docs build is an unfortunate casualty in the war on libphosh's API
surface area. Because we haven't figured out a way to build two
different GIRs for the same namespace in Meson yet, the doc build only
sees the public API and not all the internal classes.

As a result, the symbol links in docstrings to classes like
Lockshield, TopPanel, QuickSettingBox, etc must be removed.

If/when the doc build is fixed to generate based on an internal GIR
again, this commit can be reverted.

Signed-off-by: Sam Day <me@samcday.com>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1623>