media-player: Fetch cover art via http
Do that using an async method to not block the UI should the download
hang.

We avoided that so far but e.g. shortwave provides http URLs to radio
stations so lets do it and rather switch to glycene at some point.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1634>