ci: Install abi-compliance-checker in Debian image
A subsequent commit will introduce a CI job that uses this tool to
generate a report on the changes to API/ABI compatibility for the
libphosh shared library (and later, the phosh executable too).

It introduces about a half-dozen transitive dependencies and ~8mb
additional installed size. This seems acceptable and preferable to
introducing a whole additional sub-image.

Signed-off-by: Sam Day <me@samcday.com>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1639>