ci: Add abi-compliance-checker job
This job makes use of the new abi-check build option, and saves an
artifact with the new ABI dump and the HTML report that was generated
comparing it to the currently checked in baseline.

It will run automatically on tags and changes to NEWS or the libphosh
symbol list. It can also be run manually.

See: https://gitlab.gnome.org/World/Phosh/phosh/-/issues/1188
Signed-off-by: Sam Day <me@samcday.com>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1636>