build: Remove wl protos from public API altogether
This was introduced recently in 68bf9792, but as it turns out, we don't
need to leak the layer-shell-unstable wl proto header into the public
API after all. This is because the previous commit reduced the exports
of PhoshLayerSurface to just its type information.

Signed-off-by: Sam Day <me@samcday.com>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1637>