top-panel: Set padding via css
So far we set the margins on the top panel itself which was fine as long
as those were symmetric. If we set different margins this would shift
the clock out of center but setting the margins on the network and
indicator boxes doesn't work either as the case where we move the clock
off center needs the margins applied to the clock.`

So instead of worrying about all those cases use `:first-child` and
`last-child` css which handles this for us (at the expense of style
calculations on layout changes).

This also has the advantage that we don't need to worry about start and
end in RTL locales.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1655>