home: Reset overview when folded
Resetting the overview's app grid when unfolding introduces flicker when
clearing the search (when a search term is present). This is very
visible when there are running apps.

Avoid the flicker by clearing the search when the app grid isn't
visible.

As the former `reset` now mostly refreshes the thumbnails we rename it
accordingly.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1657>