background-manager: Remove background when it goes away
When we unplug a monitor the widget goes away due to the layer surface
being destroyed, make sure we remove it from the list of known
backgrounds as we otherwise might access stale data.

Instead of dropping on_monitor_removed completely we keep it as
consistency check.

That way it doesnt' matter if the widget gets destroyed before the
output or the other way around.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1665>