rotation-manager: Do not delay marking accelerometer as released
As soon as a request to release the claim on accelerometer is dispatched, mark
it as released in Phosh, even if a response from iio-sensor-proxy may only be
received much later.

This avoids the state inside Phosh getting out of sync with iio-sensor-proxy
after a suspend-resume cycle, for which the ReleaseAccelerometer response may
only be received after the system has resumed.

Closes #1214

Signed-off-by: Sicelo A. Mhlongo <absicsz@gmail.com>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1672>