gtk-mount-prompt: Use pixel-size instead of icon-size
`GtkIconSize` values are semantic in GTK 4 and no longer contain the
various values we have in GTK 3. If we need to enforce a specific
pixel-size, we need to use the `pixel-size` property of image. So let's
move to it. For the functions that need an icon-size just for namesake,
let's use `-1` so we can ensure that we have replaced its functionality
with pixel-size.

Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1682>