status-icon: Add pixel-size
`GtkIconSize` values are semantic in GTK 4 and no longer contain the
various values we have in GTK 3. If we need to enforce a specific
pixel-size, we need to use the `pixel-size` property of image. So let's
move to it.

Deprecates `icon-size`.

Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1682>