lockscreen: Check for info widgets upfront
This ensures we have a small clock e.g. when locking the screen with an
ongoing call. This will also be needed with the following commits as the
media widget will be visible initially and hence there won't be a
changed signal.

We hence move form counting invocations to checking the individual
widgets.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1692>