lockscreen: Set initial focus widget
Otherwise we end up with the focus on the first focusable widget which
is the media player's detail button and that doesn't look nice.

By setting it to `submit` we ensure that a signal tab gets it to the
first info widget.

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1692>