shell: Clean up banner correctly
The banner is a toplevel so we want to use gtk_widget_destroy on it. We
connect to the `destroy` signal so in case the layer surface invokes
`gtk_widget_destroy` we don't do it twice.

This avoids a sporadic segfault in the tests when the banner was still
being animated and also avoids a weak pointer where it is not needed.

Fixes 5d4f590b9 ("notification-banner: Add slide-up animation to banner")
(which wasn't the cause but exposed the bug).
Fixes c5556891f ("notifications: maintain a message list")

Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1704>