wall-clock: Use a regex match to get the stripped clock
`AM` and `PM` can be localized strings so look for the number part
when we have any whitespace in the clock format. Otherwise we assume
24h clock.

Hours and minutes can be separated by unicode characters so we can't
just use `:` as separator.

Closes: https://gitlab.gnome.org/World/Phosh/phosh/-/issues/1237
Signed-off-by: Guido Günther <agx@sigxcpu.org>
Part-of: <https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1731>